# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Technical Test Backend Golang",
        "title": "CAR RENTAL",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/bookingtypes": {
            "get": {
                "description": "Retrieve a list of all bookingTypes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingTypes"
                ],
                "summary": "Retrieve list of bookingTypes",
                "responses": {
                    "200": {
                        "description": "List of bookingTypes",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    },
                    "404": {
                        "description": "No bookingType found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new bookingType.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingTypes"
                ],
                "summary": "Create a new bookingType",
                "parameters": [
                    {
                        "description": "BookingType data",
                        "name": "bookingType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InputBookingType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created bookingType",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookingtypes/{id}": {
            "get": {
                "description": "Retrieve a bookingType by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingTypes"
                ],
                "summary": "Retrieve bookingType by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BookingType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BookingType data",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "BookingType not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update information of a bookingType.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingTypes"
                ],
                "summary": "Update bookingType information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BookingType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "BookingType data",
                        "name": "bookingType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InputBookingType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated bookingType",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a bookingType by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingTypes"
                ],
                "summary": "Delete bookingType by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BookingType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    },
                    "404": {
                        "description": "BookingType not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BookingType": {
            "type": "object",
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.InputBookingType": {
            "type": "object",
            "required": [
                "booking_type",
                "description"
            ],
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
        "pkg.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}